Refactor: rename app.component to vectortile-view.component
