--- viewer/src/app/app.component.html
@@ -1,5 +1,0 @@
-<div class="maparea" id="map" [style.width.px]="mapWidth" [style.height.px]="mapHeight">
-  <div class="toprightpanel">
-    <app-objectinfo-view class="objectinfopanel" id="objectinfo" [feature]="curFeature" *ngIf="showObjectInfo"> </app-objectinfo-view>
-  </div>
-</div>
