+++ viewer/src/app/feature-view/feature-view.component.ts
@@ -15,7 +15,7 @@ import { Circle, Fill, Stroke, Style } from 'ol/style'
 import WMTSTileGrid from 'ol/tilegrid/WMTS'
 import { take } from 'rxjs/operators'
 import { environment } from 'src/environments/environment'
-import { NgChanges } from '../app.component'
+import { NgChanges } from '../vectortile-view/vectortile-view.component'
 import { DataUrl, FeatureServiceService, ProjectionMapping, defaultMapping, featureCollectionGeoJSON } from '../feature-service.service'
 import { projectionSetMercator } from '../mapprojection'
 import { boxControl } from './boxcontrol'
