+++ viewer/src/app/legend-view/legend-view.component.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common'
 import { Component, ElementRef, Input, OnChanges, OnInit, ViewEncapsulation } from '@angular/core'
 import { recordStyleLayer } from 'ol-mapbox-style'
-import { NgChanges } from '../app.component'
+import { NgChanges } from '../vectortile-view/vectortile-view.component'
 import { LegendItemComponent } from '../legend-item/legend-item.component'
 import { LegendItem, MapboxStyle, MapboxStyleService } from '../mapbox-style.service'
 import { NGXLogger } from 'ngx-logger'
