+++ viewer/src/app/vectortile-view/vectortile-view.component.css
