+++ viewer/src/app/vectortile-view/vectortile-view.component.html
@@ -1,0 +1,5 @@
+<div class="maparea" id="map" [style.width.px]="mapWidth" [style.height.px]="mapHeight">
+  <div class="toprightpanel">
+    <app-objectinfo-view class="objectinfopanel" id="objectinfo" [feature]="curFeature" *ngIf="showObjectInfo"> </app-objectinfo-view>
+  </div>
+</div>
